backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
openai>=1.66.0
google-genai
requests
python-dotenv
aiofiles
pytest
pytest-asyncio
httpx
python-multipart
